# api_doc.yml
openapi: '3.0.3'
info:
  title: URL Shortening RESTful API
  version: '1.0'
  description: API documentation for URL shortening service.
  contact:
    name: Azhar Bagaskara
    email: azharbagaskara1@gmail.com
    url: https://azbagas.com
  license:
    name: APACHE 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:3000/api
    description: Local URL
paths:
  /users:
    post:
      tags:
        - users
      summary: Register new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                passwordConfirmation:
                  type: string
      responses:
        201:
          description: Register success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        409:
          description: Conflict with existing resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: Email already registered
  /users/login:
    post:
      tags:
        - users
      summary: Login user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        200:
          description: Login success. The server will include the access token in the response body and the refresh token in a cookie.
          headers:
            Set-Cookie:
              description: Set HTTP-only refresh token in a cookie for the `api/users/` path
              schema:
                type: string
                example: refreshToken=string; Path=/api/users; Secure; HttpOnly
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      accessToken:
                        type: string
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        401:
          description: Login failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: Email or password is incorrect
...
  /users/refresh:
    post:
      tags:
        - users
      summary: Get new access token
      security: []
      description:
        Retrieves a new access token using the refresh token stored in an HTTP-only cookie.
        The client must send a request with the refresh token in the cookie.
      responses:
        200:
          description: Refresh token success. The server will include the new access token in the response body.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
        401:
          description: Refresh token is invalid or expired. The user must login again.
          headers:
            Set-Cookie:
              description: Clear the refresh token in the cookie
              schema:
                type: string
                example: refreshToken=; Path=/api/users; Secure; HttpOnly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: Refresh token is invalid. Please login again.
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      description: Enter the access token from login response here
      type: http
      scheme: bearer
  schemas:
    ValidationError:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Message:
      type: object
      properties:
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        photo:
          type: string
...